+++ BuildPackages.ps1
@@ -1,13 +1,16 @@
 # =============================================
 # Build NuGet Packages for Tethys.Logging
-# SPDX-FileCopyrightText: (c) 2022-2023 T. Graf
+# SPDX-FileCopyrightText: (c) 2022-2026 T. Graf
 # SPDX-License-Identifier: Apache-2.0
 # =============================================
 
 # global version
-$version = "1.6.1"
-$copyright = "Copyright (C) 2009-2021 T. Graf"
+$version = "1.6.2"
+$copyright = "Copyright (C) 2009-2026 T. Graf"
 $all = "version=" + $version + ";copyright=" + $copyright
+
+New-Item -Path "export" -ItemType Directory -Force
+New-Item -Path "export\packages" -ItemType Directory -Force
 
 cd Tethys.Logging
 nuget pack Tethys.Logging.nuspec -properties $all
@@ -21,6 +24,11 @@ cd ..
 
 cd Tethys.Logging.Controls.NET5
 nuget pack Tethys.Logging.Controls.NET5.nuspec -properties $all
+move Tethys.Logging.*.nupkg ..\export\packages -force
+cd ..
+
+cd Tethys.Logging.Controls.NET8
+nuget pack Tethys.Logging.Controls.NET8.nuspec -properties $all
 move Tethys.Logging.*.nupkg ..\export\packages -force
 cd ..